+++ src/main/kotlin/uk/co/kievits/raytracer/together/Chapter8.kt
@@ -27,7 +27,7 @@ fun main() {
     val floor = Plane().apply {
         material.apply {
             color = Color(1, 0.9, 0.9)
-            specular = 0f
+            specular = 0.0
         }
     }
 
@@ -47,8 +47,8 @@ fun main() {
         transform = translation(-0.5, 1, 0.5)
         material.apply {
             color = Color(0.1, 1, .5)
-            diffuse = 0.7f
-            specular = 0.3f
+            diffuse = 0.7
+            specular = 0.3
         }
     }
 
@@ -56,8 +56,8 @@ fun main() {
         transform = translation(1.5, 0.5, -.5) * scaling(0.5, 0.5, 0.5)
         material.apply {
             color = Color(0.5, 1, .1)
-            diffuse = 0.7f
-            specular = 0.3f
+            diffuse = 0.7
+            specular = 0.3
         }
     }
 
@@ -65,8 +65,8 @@ fun main() {
         transform = translation(-1.5, 0.33, -.75) * scaling(.33, .33, .33)
         material.apply {
             color = Color(1, 0.8, 0.1)
-            diffuse = 0.7f
-            specular = 0.3f
+            diffuse = 0.7
+            specular = 0.3
         }
     }
 
@@ -88,7 +88,7 @@ fun main() {
     val camera = Camera(
         hSize = 800,
         vSize = 600,
-        fieldOfView = (PI / 3).toFloat(),
+        fieldOfView = PI / 3,
     )
 
     camera.transform = viewTransformation(
